+++ src/modelling/centroidal/centroidal-friction-cone.cpp
@@ -1,0 +1,8 @@
+#include "aligator/modelling/centroidal/centroidal-friction-cone.hxx"
+
+namespace aligator {
+
+template struct CentroidalFrictionConeResidualTpl<context::Scalar>;
+template struct CentroidalFrictionConeDataTpl<context::Scalar>;
+
+} // namespace aligator