+++ src/modelling/centroidal/friction-cone.cpp
@@ -1,8 +1,0 @@
-#include "aligator/modelling/centroidal/friction-cone.hxx"
-
-namespace aligator {
-
-template struct FrictionConeResidualTpl<context::Scalar>;
-template struct FrictionConeDataTpl<context::Scalar>;
-
-} // namespace aligator