+++ tests/python/test_centroidal.py
@@ -243,7 +243,7 @@ def test_friction_cone():
     mu = 0.5
     epsilon = 1e-3
 
-    fun = aligator.FrictionConeResidual(ndx, nu, k, mu, epsilon)
+    fun = aligator.CentroidalFrictionConeResidual(ndx, nu, k, mu, epsilon)
 
     fdata = fun.createData()
     fun.evaluate(x0, u0, fdata)